Added API EndPoint Interface and removed unneeded API tests.
