+++ src/main/java/com/frameworkium/core/api/Endpoint.java
@@ -1,0 +1,12 @@
+package com.frameworkium.core.api;
+
+public interface Endpoint {
+
+    /**
+     * Calls {@link String#format(String, Object...)} on the url.
+     *
+     * @param params Arguments referenced by the format specifiers in the url.
+     * @return A formatted URL String
+     */
+    String getUrl(Object... params);
+}
