+++ src/test/java/com/frameworkium/integration/tfl/api/dto/journeyplanner/Journey.java
@@ -2,8 +2,11 @@ package com.frameworkium.integration.tfl.api.dto.journeyplanner;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
+import java.util.List;
+
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class LegDto {
+public class Journey {
 
     public Integer duration;
+    public List<Leg> legs;
 }
