+++ src/test/java/com/frameworkium/integration/tfl/api/dto/journeyplanner/Leg.java
@@ -2,11 +2,8 @@ package com.frameworkium.integration.tfl.api.dto.journeyplanner;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
-import java.util.List;
-
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class JourneyDto {
+public class Leg {
 
     public Integer duration;
-    public List<LegDto> legs;
 }
