+++ src/test/java/com/frameworkium/integration/tfl/api/service/journeyplanner/ItineraryService.java
@@ -7,15 +7,9 @@ import ru.yandex.qatools.allure.annotations.Step;
 
 import java.util.Map;
 
-import static com.frameworkium.integration.tfl.api.constant.Endpoint.JOURNEY_PLANNER;
+import static com.frameworkium.integration.tfl.api.constant.TFLEndpoint.JOURNEY_PLANNER;
 
 public class ItineraryService extends BaseTFLService {
-
-    @Step("Get itinerary from {0} to {1}")
-    public Itinerary getItinerary(String from, String to) {
-        return request(JOURNEY_PLANNER.getUrl(from, to))
-                .as(Itinerary.class);
-    }
 
     @Step("Get national itinerary from {0} to {1}")
     public Itinerary getNationalItinerary(String from, String to) {
